30. Learn a simple way to authorize action

+++ app/Http/Controllers/ImageController.php
@@ -36,6 +36,10 @@ class ImageController extends Controller
 
     public function edit(Image $image): View
     {
+        if (request()->user()->id !== $image->user_id) {
+            abort(403, 'Access denied');
+        }
+
         return view('image.edit', compact('image'));
     }
 
@@ -49,6 +53,10 @@ class ImageController extends Controller
 
     public function destroy(Image $image)
     {
+        if (request()->user()->id !== $image->user_id) {
+            abort(403, 'Access denied');
+        }
+
         $image->delete();
 
         // return redirect()->route('images.index')->with('message', 'Image has been uploaded successfully');
